requirements.txt
python-binance
python-dotenv
tinydb
requests
